% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_html.R
\name{make_table_from_matrix}
\alias{make_table_from_matrix}
\title{Make table from a matrix}
\usage{
make_table_from_matrix(mat, levels)
}
\arguments{
\item{mat}{rectangular matrix}

\item{levels}{string of the levels as a JSON}
}
\value{
html list of the table
}
\description{
Make table from a matrix
}
